+++ cg_lims/models/arnold/prep/sars_cov_2_prep/pooling_and_cleanup.py
@@ -21,7 +21,7 @@ class PoolingAndCleanUpCovProcessUDFS(BaseModel):
 class PoolingAndCleanUpCovArtifactUDF(BaseModel):
     finished_library_concentration: float = Field(..., alias="Concentration")
     finished_library_concentration_nm: float = Field(..., alias="Concentration (nM)")
-    finished_library_size: Optional[float] = Field(None, alias="Size (bp)")
+    finished_library_size: Optional[int] = Field(None, alias="Size (bp)")
 
 
 class PoolingAndCleanUpCovFields(BaseStep):
